+++ apps/decodex/src/orchestrator/execution_failure/terminal_writeback/apply.rs
@@ -1,0 +1,85 @@
+use crate::{
+	orchestrator::{
+		execution_failure,
+		execution_failure::{
+			IssueRunPlan, IssueTracker, Result, TerminalFailureWritebackRuntime,
+			terminal_writeback::PreparedTerminalFailureWriteback,
+		},
+	},
+	tracker,
+};
+
+pub(crate) fn apply_terminal_failure_tracker_writeback<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	issue_run: &IssueRunPlan,
+	writeback: &PreparedTerminalFailureWriteback,
+) -> Result<()>
+where
+	T: IssueTracker,
+{
+	tracker.update_issue_state(&issue_run.issue.id, &writeback.failure_state_id)?;
+
+	apply_needs_attention_label(
+		tracker,
+		issue_run,
+		runtime.service_id,
+		&writeback.needs_attention_label,
+		writeback.needs_attention_label_id.clone(),
+		&writeback.terminal_failure_state_name,
+	)?;
+
+	if runtime.state_store.is_some() {
+		tracker::create_prepared_linear_execution_event_comment_without_remote_scan(
+			tracker,
+			&issue_run.issue.id,
+			&writeback.projection,
+		)?;
+	} else {
+		tracker::create_prepared_linear_execution_event_comment(
+			tracker,
+			&issue_run.issue.id,
+			&writeback.projection,
+		)?;
+	}
+
+	Ok(())
+}
+
+fn apply_needs_attention_label<T>(
+	tracker: &T,
+	issue_run: &IssueRunPlan,
+	service_id: &str,
+	needs_attention_label: &str,
+	needs_attention_label_id: Option<String>,
+	terminal_failure_state_name: &str,
+) -> Result<bool>
+where
+	T: IssueTracker,
+{
+	if let Some(label_id) = needs_attention_label_id.as_deref() {
+		if !tracker::issue_has_label_with_server_confirmation(
+			tracker,
+			&issue_run.issue,
+			needs_attention_label,
+		)? {
+			tracker.add_issue_labels(&issue_run.issue.id, &[label_id.to_owned()])?;
+		}
+	} else {
+		tracing::warn!(
+			label = needs_attention_label,
+			issue = issue_run.issue.identifier,
+			guard_state = terminal_failure_state_name,
+			"Needs-attention label was not found in the issue team; using a non-startable state guard when needed."
+		);
+	}
+
+	execution_failure::ensure_automation_activity_label(
+		tracker,
+		&issue_run.issue,
+		service_id,
+		false,
+	)?;
+
+	Ok(needs_attention_label_id.is_some())
+}
