+++ apps/decodex/src/orchestrator/execution_failure/terminal_writeback/prepare.rs
@@ -1,0 +1,101 @@
+use crate::{
+	orchestrator::{
+		execution_failure,
+		execution_failure::{
+			IssueRunPlan, IssueTracker, Report, Result, RetainedPartialProgress,
+			TerminalFailureLifecycle, TerminalFailureWritebackRuntime, WorkflowDocument, eyre,
+			terminal_writeback::PreparedTerminalFailureWriteback,
+		},
+	},
+	tracker,
+};
+
+pub(crate) fn prepare_terminal_failure_writeback<T>(
+	tracker: &T,
+	runtime: TerminalFailureWritebackRuntime<'_>,
+	workflow: &WorkflowDocument,
+	issue_run: &IssueRunPlan,
+	worktree_path: &str,
+	manual_attention_requested: bool,
+	error: &Report,
+) -> Result<PreparedTerminalFailureWriteback>
+where
+	T: IssueTracker,
+{
+	let tracker_policy = workflow.frontmatter().tracker();
+	let needs_attention_label = tracker_policy.needs_attention_label();
+	let needs_attention_label_id = tracker::issue_team_label_id_with_server_confirmation(
+		tracker,
+		&issue_run.issue,
+		needs_attention_label,
+	)?;
+	let failure_state_name = tracker_policy.failure_state();
+	let failure_state_is_startable =
+		tracker_policy.startable_states().iter().any(|state| state == failure_state_name);
+	let retry_guarded_by_state = needs_attention_label_id.is_none() && failure_state_is_startable;
+	let terminal_failure_state_name = if retry_guarded_by_state {
+		tracker_policy.in_progress_state()
+	} else {
+		failure_state_name
+	};
+	let failure_state_id =
+		issue_run.issue.state_id_for_name(terminal_failure_state_name).ok_or_else(|| {
+			eyre::eyre!(
+				"State `{}` was not found for issue `{}`.",
+				terminal_failure_state_name,
+				issue_run.issue.identifier
+			)
+		})?;
+	let recovery_gate = execution_failure::terminal_failure_recovery_gate(
+		needs_attention_label,
+		needs_attention_label_id.is_some(),
+		retry_guarded_by_state,
+		tracker_policy.in_progress_state(),
+	);
+	let (error_class, next_action) = execution_failure::terminal_failure_comment_details(
+		manual_attention_requested,
+		error,
+		&recovery_gate,
+	);
+	let pr_url = execution_failure::terminal_failure_pr_url(error);
+	let retained_source_error_class = error
+		.downcast_ref::<RetainedPartialProgress>()
+		.and_then(|partial_progress| partial_progress.source_error_class.as_deref());
+	let comment = execution_failure::format_terminal_failure_comment(
+		&issue_run.run_id,
+		issue_run.attempt_number,
+		worktree_path.to_owned(),
+		&issue_run.worktree.branch_name,
+		pr_url,
+		error_class,
+		&next_action,
+	);
+	let event = execution_failure::terminal_failure_lifecycle_event(
+		runtime.service_id,
+		issue_run,
+		TerminalFailureLifecycle {
+			error_class,
+			next_action: &next_action,
+			pr_url,
+			target_state: terminal_failure_state_name,
+			worktree_path,
+			manual_attention_requested,
+			retained_source_error_class,
+		},
+	);
+	let projection = tracker::prepare_linear_execution_event_comment(
+		&comment,
+		&event,
+		runtime.privacy_classifier,
+	)?;
+
+	Ok(PreparedTerminalFailureWriteback {
+		failure_state_id: failure_state_id.to_owned(),
+		needs_attention_label: needs_attention_label.to_owned(),
+		needs_attention_label_id,
+		terminal_failure_state_name: terminal_failure_state_name.to_owned(),
+		projection,
+		error_class,
+		retry_guarded_by_state,
+	})
+}
