+++ apps/decodex/src/orchestrator/run_cycle/target_issue/lease.rs
@@ -1,0 +1,38 @@
+use crate::{
+	orchestrator::run_cycle::{IssueTracker, Result, TargetIssueRunContext, eyre},
+	state::PreacquiredLeaseGuards,
+};
+
+pub(crate) fn adopt_preacquired_target_issue_lease<T>(
+	context: &TargetIssueRunContext<'_, T>,
+	issue_id: &str,
+) -> Result<()>
+where
+	T: IssueTracker,
+{
+	let preferred_run_identity = context.preferred_run_identity.ok_or_else(|| {
+		eyre::eyre!("daemon child lease handoff requires a planned run identifier")
+	})?;
+	let preferred_issue_state = context
+		.preferred_issue_state
+		.ok_or_else(|| eyre::eyre!("daemon child lease handoff requires a planned issue state"))?;
+	let issue_claim_fd = context.preferred_issue_claim_fd.ok_or_else(|| {
+		eyre::eyre!("daemon child lease handoff requires an inherited issue-claim fd")
+	})?;
+	let dispatch_slot_fd = context.preferred_dispatch_slot_fd.ok_or_else(|| {
+		eyre::eyre!("daemon child lease handoff requires an inherited dispatch-slot fd")
+	})?;
+	let dispatch_slot_index = context.preferred_dispatch_slot_index.ok_or_else(|| {
+		eyre::eyre!("daemon child lease handoff requires an inherited dispatch-slot index")
+	})?;
+
+	context.state_store.adopt_preacquired_lease(
+		context.project.service_id(),
+		issue_id,
+		preferred_run_identity.run_id,
+		preferred_issue_state,
+		PreacquiredLeaseGuards { issue_claim_fd, dispatch_slot_fd, dispatch_slot_index },
+	)?;
+
+	Ok(())
+}
